improve: Just save total & idle to calculate the cpu usage
In order to calculate the cpu usage, we don't have to save all
members in cpu stat, not to mention that these members are changing
between different kernel versions(normally kernel just adds new
members to not break userspace).

To get the cpu usage, only 2 things are required: total & idle.
So this patch changes the codes to just save these 2 things,
which simplifies the codes and also make the codes work on differnt
kernel versions.

Signed-off-by: Mark Zhang <ace119@163.com>